gstack fetched from upstream

+++ src/common/gstack.h
@@ -1,8 +1,7 @@
 /*
   gstack.h
   generic stack module
-
-  J.J.Green
+  J.J. Green
 */
 
 #ifndef GSTACK_H
@@ -12,13 +11,13 @@
 
 typedef struct gstack_t gstack_t;
 
-extern gstack_t* gstack_new(size_t,int,int);
+extern gstack_t* gstack_new(size_t, size_t, size_t);
 extern void      gstack_destroy(gstack_t*);
-extern int       gstack_push(gstack_t*,void*);
-extern int       gstack_pop(gstack_t*,void*);
+extern int       gstack_push(gstack_t*, void*);
+extern int       gstack_pop(gstack_t*, void*);
 extern int       gstack_reverse(gstack_t*);
-extern int       gstack_foreach(gstack_t*,int (*)(void*,void*),void*);
+extern int       gstack_foreach(gstack_t*, int (*)(void*, void*), void*);
 extern int       gstack_empty(gstack_t*);
-extern int       gstack_size(gstack_t*);
+extern size_t    gstack_size(gstack_t*);
 
 #endif
